arch: use u32 for initrd and image loading sizes
Change the return type of load_image() and load_image_high() from usize
to u32 for the image size to simplify address calculations and avoid
overflow handling.

Correspondingly, update create_chosen_node() and create_fdt() across all
architectures to use u32 for the initrd size.

BUG=none
TEST=presubmit

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7614527
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>

+++ aarch64/src/fdt.rs
@@ -288,7 +288,7 @@ fn create_psci_node(fdt: &mut Fdt, version: &PsciVersion) -> Result<()> {
 fn create_chosen_node(
     fdt: &mut Fdt,
     cmdline: &str,
-    initrd: Option<(GuestAddress, usize)>,
+    initrd: Option<(GuestAddress, u32)>,
     stdout_path: Option<&str>,
 ) -> Result<()> {
     let chosen_node = fdt.root_mut().subnode_mut("chosen")?;
@@ -637,7 +637,7 @@ pub fn create_fdt(
     fdt_address: GuestAddress,
     cmdline: &str,
     kernel_region: AddressRange,
-    initrd: Option<(GuestAddress, usize)>,
+    initrd: Option<(GuestAddress, u32)>,
     android_fstab: Option<File>,
     is_gicv3: bool,
     has_vgic_its: bool,
